% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_different.R
\name{all_different}
\alias{all_different}
\title{all_different}
\usage{
all_different(obj, mind)
}
\arguments{
\item{obj}{object: numeric R object which can be converted to a vector}

\item{mind}{numeric: minimum value}
}
\value{
logical
}
\description{
Test if the differences between the entries in \code{obj} is larger than \code{mind}.
}
\examples{
x <- runif(10)
all_different(x, 0.0001)
all_different(x, 1)
}
