% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_ts.R
\name{as_ts}
\alias{as_ts}
\title{as_ts}
\usage{
as_ts(ts)
}
\arguments{
\item{ts}{\code{ts_data} object}
}
\value{
\code{ts} object
}
\description{
Converts a \code{ts_data} object to a time series object (\code{ts}) in R.
}
\examples{
# time series from linear trend
ts <- ts_data(12, trend.coeff= c(sample(0:10, 1), sample(1+(1:10)/20, 1)))
as_ts(ts)
}
