% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc.R
\name{nom.cc}
\alias{nom.cc}
\alias{nom.cramer}
\alias{ord.spearman}
\alias{ord.kendall}
\title{Association}
\usage{
nom.cc(tab, correct = FALSE)

nom.cramer(tab, ...)

ord.spearman(tab, ...)

ord.kendall(tab, ...)
}
\arguments{
\item{tab}{table: contingency table with absolute frequencies}

\item{correct}{logical: if a correction should be applied (default: \code{FALSE})}

\item{...}{further parameters (unused)}
}
\value{
numeric
}
\description{
Computation of association and correlation measure:
\itemize{
\item \code{nom.cc} (corrected) contingency coefficent
\item \code{nom.cramer} Cramer's V or Phi
\item \code{ord.spearman} Spearman's rank correlation
\item \code{ord.kendall} Kendall's rank correlation
}
}
\examples{
tab <- matrix(round(10*runif(15)), ncol=5)
nom.cc(tab)
nom.cc(tab, correct=TRUE)
nom.cramer(tab)
ord.spearman(tab)
ord.kendall(tab)
}
