% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_total.R
\name{bayes_total}
\alias{bayes_total}
\title{bayes_total}
\usage{
bayes_total(target, partition = NULL, ...)
}
\arguments{
\item{target}{character: target event}

\item{partition}{character: set of events which form a partition}

\item{...}{numeric: named events with given probabilities}
}
\value{
resulting probability including the steps for computing. If \code{NA} is returned then no solution could be found.
}
\description{
Given a set of events it computes the total or conditional probability after Bayes. The events should be R names and
the available operators are
}
\details{
\itemize{
\item \code{!} complementary event,
\item \code{|} conditional event, and
\item \code{^} intersection of events.
}

The attribute \code{compute} of the return value contains the steps for computation.
}
\examples{
bayes_total('M|T', c("M", "!M"), "M"=0.6, "T|M"=0.75, "T|!M"=0.39, 0.4)
pmt <- bayes_total("M|T", c("M", "!M"), "M"=0.6, "T|M"=0.75, "T|!M"=0.39, 0.4)
cat(paste0(attr(pmt, "compute"), "\n"))
#
bayes_total('A|K', c("A", "B", "C"), "A"=0.55, "B"=0.35, "C"=0.1, "K|A"=0.4, "K|B"=0.1, "K|C"=0.1)
bayes_total('K', c("A", "B", "C"), "A"=0.55, "B"=0.35, "C"=0.1, "K|A"=0.4, "K|B"=0.1, "K|C"=0.1)
}
