% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binom_param.R
\name{binom_param}
\alias{binom_param}
\title{binom_param}
\usage{
binom_param(n, p, mean = NA, sd = NA, norm = NA, pois = NA)
}
\arguments{
\item{n}{integer: number of observations}

\item{p}{numeric: vector of probabilities}

\item{mean}{integer: number of digits the mean should have}

\item{sd}{integer: number of digits the standard deviation should have}

\item{norm}{logical: normal approximation possible}

\item{pois}{logical: poisson approximation possible}
}
\value{
a data frame with possible choices of \code{n} , \code{p}, \code{mean} and \code{sd}
}
\description{
binom_param
}
\examples{
binom_param(1000:50000, (5:25)/100, 0, 0)
}
