% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calledBy.R
\name{calledBy}
\alias{calledBy}
\title{calledBy}
\usage{
calledBy(fun = "exams2pdf")
}
\arguments{
\item{fun}{character: name of the calling function (default: \code{exams2pdf})}
}
\value{
logical
}
\description{
Checks if the result from \link[base:sys.parent]{base::sys.calls} contains a call from \code{fun}.
}
\examples{
funb <- function() { calledBy('funa') }
funa <- function() { funb() }
funa()
}
