% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.R
\name{combinatorics}
\alias{combinatorics}
\title{combinatorics}
\usage{
combinatorics(n, k)
}
\arguments{
\item{n}{total number of elements}

\item{k}{number of elements to choose}
}
\value{
a list
}
\description{
Computes all combinatorics results for k<n and returns it as list:
\describe{
\itemize{\code{permutation.n}}{\eqn{P(n)}}
\itemize{\code{permutation.k}}{\eqn{P(k)}}
\itemize{\code{permutation.nk}}{\eqn{P(n; k)}}
\itemize{\code{variation}}{\eqn{V(n;k)}}
\itemize{\code{variation.rep}}{\eqn{V^W(n;k)}}
\itemize{\code{combination}}{\eqn{K(n;k)}}
\itemize{\code{combination.rep}}{\eqn{K^W(n;k)}}
}
}
\examples{
combinatorics(4,2)
}
