% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_data.R
\name{cor_data}
\alias{cor_data}
\title{cor_data}
\usage{
cor_data(
  x,
  y,
  r,
  method = c("pearson", "kendall", "spearman"),
  ...,
  maxit = 1000
)
}
\arguments{
\item{x}{numeric: given x values}

\item{y}{numeric: given y values}

\item{r}{numeric: desired correlation}

\item{method}{character: indicating which correlation coefficient is to be computed (default: `"pearson")}

\item{...}{further parameters given to \code{\link[stats:cor]{stats::cor()}}}

\item{maxit}{numeric: maximal number of iterations (default: \code{1000})}
}
\value{
a matrix with two columns and an attribute \code{interim} for intermediate values as matrix.
The rows of the matrix contain
\itemize{
\item if \code{method=="pearson"}: \eqn{x_i}, \eqn{y_i},  \eqn{x_i-bar{x}}, \eqn{y_i-\bar{y}},
\eqn{(x_i-bar{x})^2}, \eqn{(y_i-\bar{y})^2}, and \eqn{(x_i-bar{x})((y_i-\bar{y})}.
\item if \code{method=="kendall"}:
\item if \code{method=="spearman"}: \eqn{x_i}, \eqn{y_i}, \eqn{p_i} (concordant pairs), and \eqn{q_i} (disconcordant pairs).
In a final step a vector with the row sums is appended as further column.
}
}
\description{
Generates a data set based on 'x\code{and}y\verb{for a given target correlation}r\verb{according to [stats::cor()]. The algorithm modifies the order of the}y\verb{'s, therefore is guaranteed that the (marginal) distribution of }x
and \code{y} is not modified. Note that is not guaranteed that the final correlation is the desired correlation;
the algorithm iteratively modifies the order. If you are unsatisfied with result it may help to increase \code{maxit}.
}
\examples{
x <- runif(6)
y <- runif(6)
xy <- cor_data(x, y, r=0.6)
cbind(x, y, xy)
}
