% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_n.R
\name{data_n}
\alias{data_n}
\alias{data_n25}
\alias{data_nsq}
\title{Number of observations}
\usage{
data_n(max, min = 5)

data_nsq(max, min = 5)

data_n25(max, min = 5)
}
\arguments{
\item{max}{integer: maximum sample size}

\item{min}{integer: minimum sample size (default: \code{5})}
}
\value{
a sequence of integers
}
\description{
Generates a sequence of sample sizes in a range from \code{min=5} to \code{max}
\itemize{
\item whose root is an integer (\code{data_nsq}), and
\item that are divisible only by 2 and 5 (\code{data_n25})
}
}
\examples{
data_n(10)
data_nsq(1000)
data_n25(1000)
}
