% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distribution}
\alias{distribution}
\alias{standarddistributions}
\alias{distribution.default}
\alias{quantile.distribution}
\alias{cdf}
\alias{toLatex.distribution}
\alias{is.distribution}
\title{Class distribution}
\usage{
distribution(name, ...)

standarddistributions(names = TRUE)

\method{distribution}{default}(name, ...)

\method{quantile}{distribution}(x, probs = seq(0, 1, 0.25), ...)

cdf(x, q, ...)

\method{toLatex}{distribution}(object, name = NULL, param = NULL, digits = 3, ...)

is.distribution(object, name = NULL)
}
\arguments{
\item{name}{character: a replacement of the name of the distribtuion type}

\item{...}{further named distribution parameters}

\item{names}{logical: should only the names of th standard distrutions returned or the LaTeX representations (default: \code{TRUE})}

\item{x}{distribution}

\item{probs}{numeric: vector of probabilities with values in [0,1].}

\item{q}{numeric: vector of quantiles}

\item{object}{distribution object}

\item{param}{character: names for the distrubtion parameters}

\item{digits}{integer: number of digits used in \code{signif}}
}
\value{
a distribution object
}
\description{
Holds one distribution including its parameters. The name of the distribution is used to determine, for example 
the function for quantiles: \code{paste0("q", name)}. Usually the full name has to be used; some abbreviated names are possible
\itemize{
\item{\code{binom}} hypergeometric distribution, parameters: \code{size}, \code{prob}
\item{\code{hyper}} hypergeometric distribution, parameters: \code{m}, \code{n}, \code{k}
\item{\code{geom}} geometric distribution, parameters: \code{prob}
\item{\code{pois}} Poisson distribution, parameters: \code{lambda}
\item{\code{unif}} hypergeometric distribution, parameters: \code{min}, \code{max}
\item{\code{exp}} exponential distribution, parameter: \code{rate}
\item{\code{norm}} normal distribution, parameters: \code{mean}, \code{sd}
\item{\code{lnorm}} log-normal distribution, parameters: \code{meanlog}, \code{sdlog}
\item{\code{t}} Student t distribution, parameter: \code{df}
\item{\code{chisq}} chi-squared distribution, parameter: \code{df}
\item{\code{f}} F distribution, parameters: \code{df1},  \code{df2}
}
Note that a quantile and cumulative distribution function must exist.

The following functions exists for \code{disributions}:
\itemize{
\item{\code{distribution}} creates a distribution with name `name` and parameters
\item{\code{quantile}} computes the quantiles of a distribution using `paste0('q', name)`
\item{\code{cdf}} computes the cumulative distribution function of a distribution using `paste0('p', name)`
\item{\code{is.distribution}} checks if `object` is distribution object. If `name` is given then it checks if distribution type is the same  
\item{\code{toLatex}} generates a LaTeX representation of the distribution an its parameter
}

\code{standarddistributions} returns the names of some standard distributions and their LaTeX representation
}
\examples{
d <- distribution("norm", mean=0, sd=1)
quantile(d)
quantile(d, c(0.025, 0.975))
cdf(d, 0)
is.distribution(d)
is.distribution(d, "t")
toLatex(d)
}
