% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{distributions}
\alias{distributions}
\title{Distributions}
\format{
A data frame with columns \code{r}, \code{latex}, \code{discret} and \code{package}
}
\usage{
data(distributions)
}
\description{
A data frame with the R function names, LaTeX names, discreteness and package origin of a distribution.
}
\examples{
data(distributions)
distributions
}
\keyword{datasets}
