% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equal.R
\name{equal}
\alias{equal}
\title{equal}
\usage{
equal(x, y, tol = 1e-06)
}
\arguments{
\item{x}{numeric}

\item{y}{numeric}

\item{tol}{numeric: tolerance (default: \code{1e-6})}
}
\value{
logical
}
\description{
Computes \code{abs(x-y)<tol}.
}
\examples{
equal(9*1/9, 1)
}
