% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exam2moodle_html.R
\name{exam2moodle_html}
\alias{exam2moodle_html}
\title{exam2moodle_html}
\usage{
exam2moodle_html(exam, html = NULL, pattern = ".")
}
\arguments{
\item{exam}{list: return list from \code{exam2moodle}}

\item{html}{character: name of HTML file (default: \code{NULL})}

\item{pattern}{character:  string containing a regular expression to match list elements (default: \code{.})}
}
\value{
invisibly the names of list elements in the HTML file
}
\description{
Creates a HTML page with all list elements which names match \code{pattern}. Default is to match all list elements. 
If \code{html} is \code{NULL} then no file will be written; just the name of matched list elements will be returned invisibly.
}
\examples{
x <- 1
}
