% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exams2call.R
\name{exams2call}
\alias{exams2call}
\title{exams2call}
\usage{
exams2call(prefix = "exams2")
}
\arguments{
\item{prefix}{character: start of the function name (default: \code{"exams2"})}
}
\value{
a list with the function name and its evaluated parameters
}
\description{
Returns a list with the function name and parameters called from the \code{.traceback()}. 
The function name must start with \code{"exams2"}
}
\examples{
exams2call()                 # access current call stack
}
