% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exams2moodle_html.R
\name{exams2moodle_html}
\alias{exams2moodle_html}
\title{exams2moodle_html}
\usage{
exams2moodle_html(
  exam,
  name = NULL,
  pattern = ".",
  mathjax = TRUE,
  header = 2,
  merge = list(questionlist = "<br>"),
  png = TRUE
)
}
\arguments{
\item{exam}{list: return list from \code{exam2moodle}}

\item{name}{character: name of HTML file (default: \code{NULL})}

\item{pattern}{character: string containing a regular expression to match list elements (default: \code{.})}

\item{mathjax}{logical: should be MathJax loaded (default: \code{TRUE})}

\item{header}{integer: at which level of the list a \code{<h2>...</h2>} element should be included (default: \code{2})}

\item{merge}{list: should elements with \code{.XXXXnn} at the end merged (default: \code{list('questionlist'="<br>")})}

\item{png}{logical: if a entry ends with \code{.png} then the function will try to embed the PNG in the output}
}
\value{
invisibly the names of list elements in the HTML file
}
\description{
Creates a HTML page with all list elements which names match \code{pattern}. Default is to match all list elements. 
If \code{html} is \code{NULL} then no file will be written; just the name of matched list elements will be returned invisibly.
If elements are merged then they are collapsed with \code{"\n"}. A named list of the form \code{"pattern"=collapse} 
can be used to modify the character string to separate elements, e.g. by using \code{"<br>"}.
}
\examples{
x <- 1
}
