% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exercise.R
\name{exercise}
\alias{exercise}
\alias{exercise.default}
\title{exercise}
\usage{
exercise(exer, ...)

\method{exercise}{default}(exer = NULL, ...)
}
\arguments{
\item{exer}{exercise object (default: \code{NULL})}

\item{...}{further named parameter to add to the exercise object}
}
\value{
exercise object
}
\description{
Data structure for exercise data
}
\examples{
exer <- exercise()           # new exercise
exer <- exercise(exer, x=3)  # add x to the exercise
}
