% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firstmatch.R
\name{firstmatch}
\alias{firstmatch}
\title{firstmatch}
\usage{
firstmatch(x, table, nomatch = NA_integer_)
}
\arguments{
\item{x}{character: the values to be matched; converted to a character vector if necessary}

\item{table}{character: the values to be matched against; converted to a character vector if necessary}

\item{nomatch}{integer: the value to be returned at non-matching positions (default: \code{NA_integer_})}
}
\value{
integer
}
\description{
\code{firstmatch} seeks matches for the elements of its first argument among those of its second,
for details see \code{\link[base:charmatch]{base::charmatch()}}. \code{charmatch} returns a zero if multiple matches are found, whereas \code{firstmatch}
returns the first partial match if multiple matches are found.
}
\examples{
firstmatch("d", c("chisq", "cauchy"))
charmatch("c", c("chisq", "cauchy"))
firstmatch("c", c("chisq", "cauchy"))
firstmatch("ca", c("chisq", "cauchy"))
}
