% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractions.R
\name{fractions}
\alias{fractions}
\title{fractions}
\usage{
fractions(x, cycles = 10, max.denominator = 2000, ...)
}
\arguments{
\item{x}{Any object of mode numeric. Missing values are now allowed.}

\item{cycles}{The maximum number of steps to be used in the continued fraction approximation process.}

\item{max.denominator}{An early termination criterion. If any partial denominator exceeds max.denominator the continued fraction stops at that point.}

\item{...}{arguments passed to or from other methods.}
}
\value{
An object of class "fractions". A structure with .Data component the same as the input numeric x, but with the rational approximations held as a character vector attribute, "fracs". Arithmetic operations on "fractions" objects are possible.
}
\description{
Find rational approximations to the components of a real numeric object using a standard continued fraction method.
Calls \code{\link[MASS:fractions]{MASS::fractions()}}, see details there.
}
\examples{
X <- matrix(runif(25), 5, 5)
fractions(X) #;)
fractions(solve(X, X/5))
fractions(solve(X, X/5)) + 1
}
