% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperloop.R
\name{gapply}
\alias{gapply}
\title{gapply}
\usage{
gapply(FUN, ..., .simplify = TRUE)
}
\arguments{
\item{FUN}{fucntion or character: string naming the function to be called}

\item{...}{list: arguments to the function call. The names attribute of args gives the argument names}

\item{.simplify}{logical: should the result be simplified to a data frame if possible? (default: \code{TRUE})}
}
\value{
a list or a data frame with the function results
}
\description{
Runs all combinations of elements in \code{...} as parameters of \code{FUN} (grid apply).
\code{I(.)} can be used to avoid that an element is interpreted as a grid value.
If an error occurs then the result of \code{FUN} will not be stored, you may note missing indices
in the returning list.
}
\examples{
# 8 function calls: sum(1,3,5), sum(1,3,6), ..., sum(2,4,6)
gapply("sum", 1:2, 3:4, 5:6)
# 4 function calls: sum(1,3,5:6), sum(1,4,5:66), ..., sum(2,4,5:6)
gapply("sum", 1:2, 3:4, I(5:6))
}
