% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grade.R
\name{grade}
\alias{grade}
\title{grade}
\usage{
grade(points, maxpts = max(points), fixed = TRUE)
}
\arguments{
\item{points}{numeric: points achieved in exam}

\item{maxpts}{numeric: maximal number of achievable points in exam (default: \code{max(points)})}

\item{fixed}{logical: fixed or relative grade scheme (default: \code{TRUE})}
}
\value{
grades as function of points
}
\description{
Computes a grade based on points to grade scheme by the Humboldt-Universität zu Berlin, see
§96c and §102 in the \href{https://gremien.hu-berlin.de/de/amb/2020/11/11_2020_zsp-hu_2013_ae08-2020_2020-05-29_druck.pdf#page=5}{Achte Änderung der Fächerübergreifenden Satzung zur Regelung von Zulassung, Studium und Prüfung der Humboldt-Universität zu Berlin (ZSP-HU)}
}
\examples{
x <- round(runif(100, 0, 22.4))
grade(x, 22)
}
