% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_data.R
\name{grouped_data}
\alias{grouped_data}
\title{grouped_data}
\usage{
grouped_data(x, n, compute = c("mean", "fine", "coarse"), tol = 1e-06)
}
\arguments{
\item{x}{numeric: borders}

\item{n}{numeric: absolute frequencies for each group}

\item{compute}{numeric/character: coefficint to compute}

\item{tol}{numeric: tolerance for numerical comparison}
}
\value{
a list with the class, result and a table
}
\description{
Computes mean, mode or quantile/median of grouped data.
}
\examples{
x <- 1:4
n <- ddiscrete(runif(3))
grouped_data(x, n)
}
