% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histwidth.R
\name{histwidth}
\alias{histwidth}
\title{histwidth}
\usage{
histwidth(from, to, widths, dmax = 100)
}
\arguments{
\item{from}{numeric: start value}

\item{to}{numeric: end value}

\item{widths}{numeric: a vector of width to sample from}

\item{dmax}{numeric: max. denominator value}
}
\value{
a list with `breaks`, `n`'s for each class and `decimal` if all densities are terminating decimals
}
\description{
Creates a set of breaks and absolute frequencies in the range from `from` to `to`. The class widths are 
sampled from `widths`. The resulting numbers could be multiplied with an integer if the `sum(n)` is too small.
Additionally is checked if the densities generated are terminating decimals.
}
\examples{
l <- histwidth(1.6, 2.1, widths=c(0.05, 0.1, 0.15, 0.2))
l
x <- histx(l$breaks, l$n)
histdata(x, l$breaks)
}
