% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_e2m.R
\name{html_e2m}
\alias{html_e2m}
\title{html_e2m}
\usage{
html_e2m(
  exam,
  name = NULL,
  pattern = ".",
  mathjax = TRUE,
  browseURL = TRUE,
  overwrite = FALSE,
  header = 2,
  merge = list(questionlist = "<br>"),
  png = TRUE
)
}
\arguments{
\item{exam}{list: return list from \code{exam2moodle}}

\item{name}{character: name of HTML file (default: \code{NULL})}

\item{pattern}{character: string containing a regular expression to match list elements (default: \code{.})}

\item{mathjax}{logical: should MathJax be loaded (default: \code{TRUE})}

\item{browseURL}{logical: should the generated HTML shown (default: \code{TRUE})}

\item{overwrite}{logical: should the HTML file overwritten, if it exists (default: \code{FALSE})}

\item{header}{integer: at which level of the list a \code{<h2>...</h2>} element should be included (default: \code{2})}

\item{merge}{list: should elements with \code{.XXXXnn} at the end merged (default: \code{list('questionlist'="<br>")})}

\item{png}{logical: if a entry ends with \code{.png} then the function will try to embed the PNG in the output}
}
\value{
invisibly the names of list elements in the HTML file
}
\description{
Creates an HTML page with all the contents of XML tags whose names match \code{pattern}.\cr\cr
The default is to show the contents of all XML tags. The HTML page is stored in the HTML file \code{name}.\cr\cr
The default \code{name=NULL} creates a temporary file. If the name does not end in code{.html}, then a \code{.html} is appended.

If \code{browseURL=TRUE} (default) then the HTML page will be displayed in the browser.

If necessary the contents of XML tags are concatenated with \code{"\n"}.
For single XML tags this can be changed, e.g. \code{merge=list("questionlist"="<br>"} leads to the fact that for the XML tag
\code{<questionlist>...</questionlist>)} \code{"<br>"} is used instead of \code{"\n"}.
}
\examples{
if (interactive()) {
  resexams <- readRDS(system.file("xml", "klausur-test.rds", package="exams.moodle"))
  html_e2m(resexams) # opens HTML file into browser
}
}
\seealso{
The aim is similar to \code{exams:::exams:::browse_exercise}, but \code{html_e2m} takes the information form
the XML file generated by the \code{exams2moodle}.
}
