% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperloop.R
\name{hyperloop}
\alias{hyperloop}
\title{hyperloop}
\usage{
hyperloop(FUN, ...)
}
\arguments{
\item{FUN}{function with named parameter(s)}

\item{...}{named parameters which contain lists with possible parameter values}
}
\value{
a hyperloop object
}
\description{
Runs a function several times with all parameter combinations:
}
\details{
\itemize{
\item If an argument is not a list then it will be converted to a one element list.
\item If an error occurs then the result of \code{FUN} will not be stored-
}
}
\examples{
x   <- rnorm(100)
trm <- hyperloop(mean, x=list(x), trim=as.list(seq(0, 0.5, by=0.05)))
# automatic conversion of x to list(x)
trm <- hyperloop(mean, x=x, trim=as.list(seq(0, 0.5, by=0.05))) 
unlist(trm)
}
