% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.prob.R
\name{is.prob}
\alias{is.prob}
\title{is.prob}
\usage{
is.prob(x, open = TRUE, min = 0, max = 1)
}
\arguments{
\item{x}{numeric: values to check}

\item{open}{logical(1): are the left and right borders are open or closed (default: \code{TRUE})}

\item{min}{numeric: minimal value (default: \code{0})}

\item{max}{numeric: maximal value (default: \code{1})}
}
\value{
a logical vector with the same length as \code{x}
}
\description{
Checks if \code{x} is in an opened/closed interval between \code{min} and \code{max}.
The defaults are choosen such that the interval is an the interval \eqn{(0,1)}, for example if \code{x} is a probability.
}
\examples{
is.prob(runif(1))
}
