% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.tag.R
\name{is.tag}
\alias{is.tag}
\title{is.tag}
\usage{
is.tag(xml, tagname = NULL)
}
\arguments{
\item{xml}{list: xml list with tags and bodies}

\item{tagname}{character: vector of tag names (default: \code{NULL})}
}
\value{
a logical vector
}
\description{
Returns a logical vector if the entry in \code{xml} is a tag or body. If \code{tagname} is given then 
it returns \code{TRUE} if the tag name appears in \code{tagname}.
}
\examples{
file <- system.file('xml', 'rexams.xml', package="exams2moodle")
xml  <- moodle_xml2list(file, FALSE)
# position body entries 
which(!is.tag(xml))
# position of question tags
which(is.tag(xml, "question"))
}
