% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_terminal.R
\name{is_terminal}
\alias{is_terminal}
\title{is_terminal}
\usage{
is_terminal(x)
}
\arguments{
\item{x}{numeric: values to test}
}
\value{
logical
}
\description{
Checks whether all \code{x}'s can be expressed as a terminal fraction,
which means the denominators are only dividable by two and five.
}
\examples{
is_terminal(2/3) # 0.6666... non-terminal
is_terminal(1/5) # 0.2       terminal
}
