% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{latexdef}
\alias{latexdef}
\alias{answercol}
\title{latexdef}
\usage{
latexdef(name, body)

answercol(n)
}
\arguments{
\item{name}{character: macro name}

\item{body, n}{character: macro body}
}
\value{
nothing
}
\description{
If exams is called by \code{exams2pdf}
\itemize{
\item \code{latex1def} adds a \verb{\\def\\name\{body\}}, and
\item \code{answercol} adds a \verb{\\def\\answercol\{n\}}
to the LaTeX file.
}
}
\examples{
answercol(2)
}
