% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcmval.R
\name{lcmval}
\alias{lcmval}
\title{lcmval}
\usage{
lcmval(x)
}
\arguments{
\item{x}{integer: numbers to find the least common multiple}
}
\value{
the least common multiple
}
\description{
Computes the least common multiple for a numeric vector \code{x}.
}
\examples{
lcmval(c(144, 160))      # = 1440
lcmval(c(144, 160, 175)) # = 50.400
}
