% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makekey.R
\name{makekey}
\alias{makekey}
\title{makekey}
\usage{
makekey(index)
}
\arguments{
\item{index}{integer: vector of integer}
}
\value{
a character
}
\description{
generates a character key from a vector of integers
}
\examples{
makekey(1)
makekey(1:2)
makekey(pi) # ;)
makekey(c(5,4))
}
