% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcval.R
\name{mcval}
\alias{mcval}
\alias{mcval.default}
\alias{mcval.histogram}
\title{mcval}
\usage{
mcval(x, ...)

\method{mcval}{default}(x, ...)

\method{mcval}{histogram}(x, exact = FALSE, ...)
}
\arguments{
\item{x}{data object}

\item{...}{unused}

\item{exact}{logical: compute exact mode or use class mids (default: \code{FALSE})}
}
\value{
a vector of modes
}
\description{
Computes all modes (most common value).
}
\examples{
x <- sample(1:5, 15, replace=TRUE)
mcval(x)
}
