% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanint_data.R
\name{meanint_data}
\alias{meanint_data}
\title{meanint_data}
\usage{
meanint_data(x, r = range(x), ...)
}
\arguments{
\item{x}{numeric: number of observations or x values}

\item{r}{numeric(2): range where the x values allowed (default: \code{range(x)})}

\item{...}{further parameters given to \code{mean}}
}
\value{
a set of integer observations with an integer mean
}
\description{
meanint_data
}
\examples{
x <- meanint_data(10, c(1, 10))
mean(x)
}
