% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polynomial.R
\name{monomial}
\alias{monomial}
\title{monomial}
\usage{
monomial(degree = 1, coefficient = 1)
}
\arguments{
\item{degree}{integer: degree of the polynomial (default: \code{1})}

\item{coefficient}{numeric: coefficient of the polynomial (default: \code{1})}
}
\value{
a poylnomial
}
\description{
Creates a polynomial of the form \eqn{c*x^d}.
}
\examples{
monomial()     # equivalent to polynomial()
monomial(3)    # x^3
monomial(3, 2) # 2*x^3
}
