% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle_fraction.R
\name{moodle_fraction}
\alias{moodle_fraction}
\title{moodle_fraction}
\usage{
moodle_fraction(xml, newfraction = NULL)
}
\arguments{
\item{xml}{list: list of XML tags and bodies}

\item{newfraction}{list: fraction values (default: \code{NULL})}
}
\value{
If \code{newfraction} is \code{NULL} then it returns a list with fractions.
        Otherwise it returns the modified XML list.
}
\description{
Extracts and sets from the XML list all fraction attributes from question tags.
If \code{newfraction} is \code{NULL} then list all fraction attributes
is created and returned separately for each question. Otherwise the entries
in \code{newfraction} replace the ones in the XML list.
}
\examples{
file <- system.file('xml', 'rexams.xml', package="exams2moodle")
xml  <- moodle_xml2list(file, FALSE)
fraction <- moodle_fraction(xml)
fraction[[2]]
fraction[[2]] <- c(0,0,100,100,0)
xml <- moodle_fraction(xml, fraction)
}
