% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle_fractions.R
\name{moodle_fractions}
\alias{moodle_fractions}
\title{moodle_fraction}
\usage{
moodle_fractions(all = TRUE)
}
\arguments{
\item{all}{logical: return all fractions or only positive and zero fractions (default: \code{TRUE})}
}
\value{
a numeric vector
}
\description{
Returns a vector with all possible Moodle fractions, see also \code{exams:::moodleFractions}
}
\examples{
moodle_fractions()
moodle_fractions(FALSE)
}
