% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle_list2xml.R
\name{moodle_list2xml}
\alias{moodle_list2xml}
\title{moodle_list2xml}
\usage{
moodle_list2xml(taglist, file, quiet = FALSE)
}
\arguments{
\item{taglist}{list: list of XML tags and bodies}

\item{file}{character: XML file name}

\item{quiet}{logical: generate output during writing (default: \code{FALSE})}
}
\value{
nothing
}
\description{
Writes a XML list, read with \code{moodle_xml2list}, to a file.
}
\examples{
file <- system.file('xml', 'rexams.xml', package="exams2moodle")
xml  <- moodle_xml2list(file, FALSE)
tmpfile <- tempfile()
moodle_list2xml(xml, tmpfile)
# now compare both files, except for some empty lines there should be no difference  
}
