% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle_m2s.R
\name{moodle_m2s}
\alias{moodle_m2s}
\title{moodle_m2s}
\usage{
moodle_m2s(file, newfile = NULL, verbose = 1)
}
\arguments{
\item{file}{character: Moodle XML file with exercises to read from}

\item{newfile}{character:  Moodle XML file to write to (default: \code{file})}

\item{verbose}{integer: generate output (default: \code{1})}
}
\value{
invisibly the file name written to
}
\description{
The \code{exams} package does not support multiple-choice questions with multiple correct answers but only one allowed answer;
but Moodle does. The function reads the XML file generated by \code{exams2moodle} and changes for \strong{all} \code{mchoice} questions:
\itemize{
\item \code{<single>...</single>} to \code{<single>true</single>}, and
\item modifies the attribute \code{fraction} in the tags \code{<answer fraction="...">...</answer>}.
If \code{fraction} is less than 0, it is set to zero, and if \code{fraction} is greater than 0, it is set to 100.
}

If the \code{file} does not end with \code{.xml}, \code{.xml} is appended. At the end, the modified XML code is stored in \code{newfile}.
}
\examples{
if (interactive()) {
  newfile <- tempfile(fileext=".xml")
  moodle_m2s(system.file("xml", "klausur-test.xml", package="exams.moodle"), newfile=newfile)
  file.edit(newfile)
}
}
