% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle_text.R
\name{moodle_text}
\alias{moodle_text}
\title{moodle_text}
\usage{
moodle_text(xml)
}
\arguments{
\item{xml}{list: a list with XML tags and bodys}
}
\value{
list with questions, feedbacks and further informations
}
\description{
Extracts information from the XML list about questions.
}
\examples{
x <- 1
}
