% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle_text2html.R
\name{moodle_text2html}
\alias{moodle_text2html}
\title{moodle_text2html}
\usage{
moodle_text2html(xml, html = sub(".xml$", "_text.html", xml), elem = NULL)
}
\arguments{
\item{xml}{character: name of XML file}

\item{html}{character: name of HTML file (default: \code{sub(".xml$", "_text.html", xml)} )}

\item{elem}{character: text elements to put in the HTML file  (default: \code{NULL} = all elements)}
}
\value{
invisibley the HTML file name
}
\description{
Extracts certain \code{<text>...</text>} elements from the XML file and creates an HTML file from it.
}
\examples{
x <- 1
}
