% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle_xml2list.R
\name{moodle_xml2list}
\alias{moodle_xml2list}
\title{moodle_xml2list}
\usage{
moodle_xml2list(file, quiet = TRUE)
}
\arguments{
\item{file}{character: XML file name}

\item{quiet}{logical: generate output during reading (default: \code{FALSE})}
}
\value{
a list with XML tags and bodys
}
\description{
Reads in a XML Moodle file generated with \code{exams2moodle} from the \code{exams} package.
}
\examples{
# Read in a XML file generated from the test exercise from the exams package
file <- system.file('xml', 'rexams.xml', package="exams2moodle")
xml  <- moodle_xml2list(file, FALSE)
head(xml)
}
