% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearest_arg.R
\name{nearest_arg}
\alias{nearest_arg}
\title{nearest_arg}
\usage{
nearest_arg(arg, choices, method = "cosine", ...)
}
\arguments{
\item{arg}{character: vector or NULL}

\item{choices}{character: vector of candidate values}

\item{method}{character: method for distance calculation (default: \code{cosine})}

\item{...}{further parameters for \link[stringdist:stringdist]{stringdist::stringdistmatrix}}
}
\value{
for each value in \code{arg} the (first) nearest element of \code{choices}
}
\description{
It determines the nearest candidate value for each value in \code{arg}. As a replacement for \link[base:match.arg]{base::match.arg}
since it is more (typo) error tolerant, but if a wrong choice will be done it is difficult to detect.
}
\examples{
# match.arg("tow.sided", c("two.sided", "less", "greater")) # will fail
nearest_arg("tow.sided", c("two.sided", "less", "greater")) 
nearest_arg(c("two.sided", "less", "greater"), c("two.sided", "less", "greater"))
nearest_arg(c("two", "two", "ded", "ss", "ea"), c("two.sided", "less", "greater"))
}
