% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nosanitize.R
\name{nosanitize}
\alias{nosanitize}
\title{Sanitation functions}
\usage{
nosanitize(str)
}
\arguments{
\item{str}{character: vector to sanitize}
}
\value{
sanitized cahracter vector
}
\description{
\code{nosanitize} makes no sanitation on the strings.
}
\examples{
nosanitize("Test")
}
