% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/now.R
\name{now}
\alias{now}
\title{now}
\usage{
now(last = 35)
}
\arguments{
\item{last}{integer: how many digits should be returned (default: \code{35})}
}
\value{
a character
}
\description{
Returns a time stamp based on the current time. 
If the clock resolution is not small enough then the function will throw an error.
}
\examples{
now()   # return all digits
now(3)  # return only the first three digits
}
