% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsprintf.R
\name{nsprintf}
\alias{nsprintf}
\alias{round_de}
\alias{schoice_de}
\title{nsprintf, round_de}
\usage{
nsprintf(n, ...)

round_de(n)

schoice_de()
}
\arguments{
\item{n}{integer: number(s) to be used}

\item{...}{character: format strings to be used}
}
\value{
\code{sprintf}ed strings
}
\description{
\code{nsprintf} creates a text dependent on the value(s) in \code{n}. Specially, we have
\itemize{
\item \code{round_de} returns either \verb{Runden Sie Ihr Ergebnis auf eine ganze Zahl},
\verb{Runden Sie Ihr Ergebnis auf eine Stelle nach dem Komma},  or
\verb{Runden Sie Ihr Ergebnis auf n Stellen nach dem Komma}
\item \code{schoice_de} returns \verb{Es kann eine oder mehrere Antworten richtig sein. Es ist ausreichend, eine richtige Antwort anzugeben. Geben Sie mehrere Antworten an und eine ist falsch, dann ist die Aufgabe falsch beantwortet}
}
}
\examples{
nsprintf(0, '0' = "keine Netzunterbrechung", '1' = "eine Netzunterbrechung", 
            "\%i Netzunterbrechungen")
nsprintf(0:3, `0` = "keine Netzunterbrechung", `1` = "eine Netzunterbrechung", 
              "\%i Netzunterbrechungen")
}
