% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num2str.R
\name{num2str}
\alias{num2str}
\title{num2str}
\usage{
num2str(..., denom = -1)
}
\arguments{
\item{...}{numeric variables}

\item{denom}{integer: denominator for fractional number}
}
\value{
a list
}
\description{
Converts a set of numeric variables to a list as string representation, either as decimal or fractional numebr.
}
\examples{
x <- 1
l <- num2str(x)         # returns in l$x the string representation
l <- num2str(x, y=x+1)  # returns in l$x and l$y the string representations
}
