% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_result.R
\name{num_result}
\alias{num_result}
\title{num_result}
\usage{
num_result(x, digits = NULL, tolerance = NULL, tolmult = 1, ...)
}
\arguments{
\item{x}{numeric: rounded data}

\item{digits}{numeric: number of digits to rounding (default: \code{NULL})}

\item{tolerance}{numeric: tolerance for rounded data (default: \code{NULL})}

\item{tolmult}{numeric: mutiplier for tolerance}

\item{...}{further parameters to \code{\link[exams:fmt]{exams::fmt()}}}
}
\value{
a list
}
\description{
Creates a list with the elements
}
\details{
\itemize{
\item \code{x} the original values
\item \code{fx} the rounded values with \code{\link[exams:fmt]{exams::fmt()}} as character
\item \code{tolerance} the tolerance
\item \code{digits} the digits used for rounding
}
}
\examples{
# height for german man (in meter)
x <- rnorm(10, mean=1.8, sd =0.25)
num_result(c(mean(x), x), digits=2)
}
