% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NumberProperties.R
\name{divisor_25}
\alias{divisor_25}
\alias{is_terminal}
\alias{prime_numbers}
\alias{primes}
\title{divisor_25, is_terminal}
\usage{
divisor_25(x)

is_terminal(x)

prime_numbers(n, sieve = FALSE)

primes(x, min = 2)
}
\arguments{
\item{x}{numeric: values to test/check}

\item{n}{integer: find all prime numbers up to n}

\item{sieve}{logical: should in any case the Sieve of Eratosthenes be used to compute prime numbers (default: \code{FALSE})}

\item{min}{integer: minimal prime number used (default: \code{2})}
}
\value{
logical
}
\description{
\itemize{
\item \code{is_terminal} checks whether all \code{x}'s can be expressed as a terminal fraction, basically \verb{all(divisor_25(denominator(x))))}
\item \code{divisor_25} checks all \code{x}'s can be expressed as \eqn{2^x 5^y}
\item \code{prime_numbers} returns all prime numbers up to a limit
\item \code{primes} prime factorization of \code{x}, returns a matrix with the power for each prime number
}
}
\examples{
is_terminal(2/3)   # 0.6666... non-terminal
is_terminal(1/5)   # 0.2       terminal
divisor_25(1:25)
prime_numbers(100) # all prime numbers less equal 100
primes(1:20)       # prime factorization of 1 to twenty
}
