% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pintegrate.R
\name{pintegrate}
\alias{pintegrate}
\title{pintegrate}
\usage{
pintegrate(pcoeff, interval, power = 0)
}
\arguments{
\item{pcoeff}{numeric: vector with coefficients for polynom, last entry is the intercept}

\item{interval}{numeric: interval to integrate over}

\item{power}{numeric: power of x to multiply the polynomial (default: \code{power=0})}
}
\value{
the definite integral(s)
}
\description{
Computes the definite integral for a polynomial function of the form (\code{q=length(pcoeff)-1}):
\deqn{\int_{\min(interval)}^{\max(interval)} (pcoeff[1] x^q + pcoeff[2] x^{q-1} + ... + pcoeff[q+1]) x^{power} dx}.
}
\examples{
# poly(x) = -0.5 x + 1 
res <- pintegrate(c(-0.5, 1), c(0, 2), 0:2)
# is density
res[1]
# expectation
res[2]
# variance
res[3]-res[2]^2
}
