% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/q2norm.R
\name{q2norm}
\alias{q2norm}
\title{q2norm}
\usage{
q2norm(x, probs = c(0.025, 0.975))
}
\arguments{
\item{x}{numeric(2): the quantiles}

\item{probs}{numeric(2): probabilities with values in \eqn{[0,1]} (default: \code{c(0.025, 0.975)})}
}
\value{
a list with a component \code{mean} and \code{sd}
}
\description{
Given two (or more) quantiles it computes a (approximate) mean and standard deviation for a corresponding normal distribution.
}
\examples{
q2norm(c(100,200))
}
