% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions.R
\name{questions}
\alias{questions}
\title{questions}
\usage{
questions(html, template = NULL, yaml = NULL, nowlen = 4)
}
\arguments{
\item{html}{list: result from \link{moodle_text}}

\item{template}{character: file name (default: \code{NULL})}

\item{yaml}{list: list with yaml parameters (default: \code{NULL})}

\item{nowlen}{integer: length of time stamps to search for}
}
\value{
a list with RMarkdown documents
}
\description{
Based on the results of  \link{moodle_text} and a template file a RMarkdown file is created for each exercise.
If a template is not given then \code{system.file('template', 'question_german', package="exams2moodle")} is used.
If no yaml is given then no YAML header will be inserted into the generated Rmd files.
}
\examples{
x <- 1
}
