% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{random}
\alias{random}
\title{random}
\usage{
random(v)
}
\arguments{
\item{v}{vector: vector with elements}
}
\value{
index
}
\description{
Returns a index from \code{1:length(v)} randomly ordered.
}
\examples{
random(-3:3)
}
