% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_fmt.R
\name{replace_fmt}
\alias{replace_fmt}
\title{replace}
\usage{
replace_fmt(txt, digits = 2L, ...)
}
\arguments{
\item{txt}{character: text where replacment is done}

\item{digits}{numeric or list: number of digits to round}

\item{...}{names to replace with values}
}
\value{
character with replaced names
}
\description{
Replaces in a text names with 
* values which are formatted with [exams::fmt()], or
* strings.
}
\examples{
replace_fmt("\\\\frac{x}{y}", x=2, y=3)
replace_fmt("\\\\frac{x}{y}", x=2, y=3, digits=0)
replace_fmt("\\\\frac{x}{y}", x=2, y=3, digits=list(0))
replace_fmt("\\\\frac{x}{y}", x=2, y=3, digits=list(2, y=0))
replace_fmt("\\\\frac{x}{y}", x="\\\\\\\\sum_{i=1}^n x_i", y="\\\\\\\\sum_{i=1}^n y_i")
}
