% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_size.R
\name{sample_size_freq}
\alias{sample_size_freq}
\title{sample_size_freq}
\usage{
sample_size_freq(n, f, which = NA)
}
\arguments{
\item{n}{numeric: vector of sample size(s) to check}

\item{f}{numeric: vector of relative frequencies}

\item{which}{numeric: if several \code{n}s are possible then return the \code{which}th one (default: \code{NA} = choose randomly one)}
}
\value{
one sample size
}
\description{
Checks if a vector of possible sample sizes and relative frequencies create integer absolute frequencies.
}
\examples{
f <- ddiscrete(runif(5), unit=100)
sample_size_freq(seq(10, 200, 1), f)
sample_size_freq(seq(10, 200, 1), f, which=200)
}
