% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{skalenniveau}
\alias{skalenniveau}
\title{skalenniveau}
\format{
A data frame with columns \code{var}, and \code{type}
}
\usage{
data(skalenniveau)
}
\description{
A data frame with the variables and level of measurement type. The names are in German.
}
\examples{
data(skalenniveau)
head(skalenniveau)
}
\keyword{datasets}
