% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sos100}
\alias{sos100}
\alias{sos200}
\alias{sos400}
\alias{sos800}
\alias{sos1000}
\title{Precomputed sum of squares data}
\format{
For each line of a matrix holds \eqn{\sum_{i=1}^k x_i^2=n} and \eqn{\sum_{i=1}^k x_i=0}. 
It contains all integer solutions up to \code{k<=10}. A \code{NA} means that this entry is not used.

An object of class \code{matrix} (inherits from \code{array}) with 2433 rows and 10 columns.

An object of class \code{matrix} (inherits from \code{array}) with 15533 rows and 10 columns.

An object of class \code{matrix} (inherits from \code{array}) with 118696 rows and 10 columns.

An object of class \code{matrix} (inherits from \code{array}) with 229830 rows and 10 columns.
}
\usage{
data(sos)

sos200

sos400

sos800

sos1000
}
\description{
Five data matrices with precomputed results from \code{sumofsquares(n, 10, zerosum=TRUE, maxt=Inf)} for
\code{n=100}, \code{n=200}, \code{n=400},\code{n=800}, and \code{n=1000}.
}
\examples{
data(sos)
head(sos100)
rowSums(sos100^2, na.rm=TRUE)
rowSums(sos100, na.rm=TRUE)
}
\keyword{datasets}
