% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{template}
\alias{template}
\title{template}
\usage{
template(tmpl, ...)
}
\arguments{
\item{tmpl}{character: template}

\item{...}{named parameter used in the template}
}
\value{
character where the R code is replaced by its evaluation
}
\description{
A text template where R code can be embedded.
}
\examples{
tmpl <- "`r a`+`r b`"
template(tmpl, a=1, b=2)
}
