% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.R
\name{tooltip}
\alias{tooltip}
\title{tooltip}
\usage{
tooltip(x, tooltip = NULL)
}
\arguments{
\item{x}{htlm_matrix object}

\item{tooltip}{character: text to show (default: \code{NULL})}
}
\value{
html_matrix object
}
\description{
Adds a text tooltip to the HTML matrix.
}
\examples{
library("magrittr")
library("tools")
m    <- matrix(1:12, ncol=4)
hm   <- html_matrix(m) \%>\% 
          tooltip(sprintf("Table has \%0.f rows and \%0.f columns", nrow(.), ncol(.)))
html <- toHTML(hm, browser=TRUE)
}
