% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec2mat.R
\name{vec2mat}
\alias{vec2mat}
\title{vec2mat}
\usage{
vec2mat(x, colnames = NULL, rownames = NULL, horizontal = TRUE)
}
\arguments{
\item{x}{vector:}

\item{colnames}{character: vector of new column names (default: \code{NULL})}

\item{rownames}{character: vector of new row names (default: \code{NULL})}

\item{horizontal}{logical: horizontal or vertical matrix (default: \code{TRUE})}
}
\value{
matrix
}
\description{
Converts a vector horizontal or vertical matrix and sets row- or colnames. 
If \code{rownames} or \code{colnames} are given then existing row names or column names are overwritten.
}
\examples{
x <- runif(5)
vec2mat(x)
vec2mat(x, horizontal=FALSE)
}
