% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warningc.R
\name{warningc}
\alias{warningc}
\title{warningc}
\usage{
warningc(...)
}
\arguments{
\item{...}{further parameters (unused)}
}
\value{
nothing
}
\description{
Compactified warning messages.
}
\examples{
test <- function() {
  warning("Warning 1")
  warning("Warning 1")
  warning("Warning 2")
}
test()
warningc()
}
