% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zebra.R
\name{zebra}
\alias{zebra}
\title{zebra}
\usage{
zebra(x, col = c("#FFFFFF", "#CCCCCC"), byrow = TRUE)
}
\arguments{
\item{x}{html_matrix object}

\item{col}{a vector of colors to zebra with (default:\code{c("#FFFFFF", "#CCCCCC")})}

\item{byrow}{logical: zebra by row or by column (default: \code{TRUE})}
}
\value{
html_matrix object
}
\description{
zebra
}
\examples{
library("magrittr")
library("tools")
m    <- matrix(1:12, ncol=4)
hm   <- html_matrix(m) \%>\% zebra()
html <- toHTML(hm, browser=TRUE)
}
