% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_string.R
\name{as_string}
\alias{as_string}
\title{as_string
converts a character vector in a single string}
\usage{
as_string(txt, collapse = ", ", last = ", and ")
}
\arguments{
\item{txt}{vector: (character) vector to merge}

\item{collapse}{character: glue text between elements (default: \code{", "})}

\item{last}{character: glue text between the two last elements  (default: \code{", and "})}
}
\value{
a string
}
\description{
as_string
converts a character vector in a single string
}
\examples{
x <- runif(5)
y <- c(TRUE, FALSE, NA) 
as_string(x)
as_string(y)
# toString
as_string(as.character(x))
as_string(as.character(y))
}
