% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.R
\name{combinatorics}
\alias{combinatorics}
\alias{variation}
\alias{combination}
\alias{permutation}
\title{Combinatorics}
\usage{
combinatorics(n, k)

variation(n, k, repl = FALSE)

combination(n, k, repl = FALSE)

permutation(n, k = rep(1, n))
}
\arguments{
\item{n}{numeric: total number of elements}

\item{k}{numeric: number of elements to choose}

\item{repl}{logical: with reptition (default: \code{FALSE})}
}
\value{
a list
}
\description{
\itemize{
\item \code{permutation} computes the number of permutations
\item \code{variation} computes the number of variations with and without replication
\item \code{combination} computes the number of combinatios with and without replication
\item \code{combinatorics} computes all combinatorics results for k<n and returns it as list:
\describe{
\itemize{\code{permutation.n}}{\eqn{P(n)}}
\itemize{\code{permutation.k}}{\eqn{P(k)}}
\itemize{\code{permutation.nk}}{\eqn{P(n; k)}}
\itemize{\code{variation}}{\eqn{V(n;k)}}
\itemize{\code{variation.rep}}{\eqn{V^W(n;k)}}
\itemize{\code{combination}}{\eqn{K(n;k)}}
\itemize{\code{combination.rep}}{\eqn{K^W(n;k)}}
}
}
}
\examples{
permutation(8)
permutation(8, c(1,3,2,2))
combination(8, 4)
combination(8, 4, TRUE)
variation(8, 4)
variation(8, 4, TRUE)
combinatorics(8, 4)
}
