% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddiscrete2.R
\name{ddiscrete2}
\alias{ddiscrete2}
\title{ddiscrete}
\usage{
ddiscrete2(n, unit = NULL, zero = FALSE)
}
\arguments{
\item{n}{numeric(2): rows and columns of the joint probability}

\item{unit}{integer: reciprocal of smallest non-zero probability (default: \code{NULL})}

\item{zero}{logical: zeros in the final probabilities allowed (default: \code{FALSE})}
}
\value{
a discrete probability function
}
\description{
Creates a bivariate discrete probability function based on \code{x} with resolution \code{unit}. 
If \code{unit} is not given then unit will be 10, 100, 1000, ... depending on the length of 
the discrete probability function.
}
\examples{
ddiscrete2(c(2,3))
ddiscrete2(c(2,3), unit=100)
ddiscrete2(c(2,3), unit=10, zero=TRUE)
}
