% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exams2moodle_html.R
\name{exams2moodle_html}
\alias{exams2moodle_html}
\title{exams2moodle_html}
\usage{
exams2moodle_html(exam, name = NULL, pattern = ".")
}
\arguments{
\item{exam}{list: return list from \code{exam2moodle}}

\item{name}{character: name of HTML file (default: \code{NULL})}

\item{pattern}{character: string containing a regular expression to match list elements (default: \code{.})}
}
\value{
invisibly the names of list elements in the HTML file
}
\description{
Creates a HTML page with all list elements which names match \code{pattern}. Default is to match all list elements. 
If \code{html} is \code{NULL} then no file will be written; just the name of matched list elements will be returned invisibly.
}
\examples{
x <- 1
}
