% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html_matrix.R
\name{html_matrix}
\alias{html_matrix}
\title{html_matrix}
\usage{
html_matrix(
  x,
  title = NULL,
  caption = NULL,
  format = list(),
  style = list(),
  names = list(),
  type = "html"
)
}
\arguments{
\item{x}{matrix: matrix to print in HTML}

\item{title}{character: matrix title, the left upper entry of the matrix (default: \code{NULL})}

\item{caption}{character: caption (default: \code{NULL})}

\item{format}{list: list of parameters to format matrix entries (default: \code{list()})}

\item{style}{list: list of parameters to set the style parameters for each matrix entry (default: \code{list()})}

\item{names}{list: list of row- and/or colnames (default: \code{list()})}

\item{type}{character: output format, either \code{"html"}, \code{"wiki"} or \code{"plain"}  (default: \code{"html"})}
}
\value{
Plain text, HTML or MediaWiki code
}
\description{
Returns a HTML representation of a matrix with the following parameters and defaults (Note that recycling can used with most parameters):
\describe{
\item{\code{title}}{entry at the top left (default: \code{""})}
\item{\code{caption}}{entry for the caption (default: \code{""})}
\item{\code{names$col}}{entry for the column names (default: \code{colnames(x)})}
\item{\code{names$row}}{entry for the row names (default: \code{rownames(x)})}
\item{\code{style$table}}{style for the table (default: \code{""})}
\item{\code{style$caption}}{style for the caption (default: \code{""})}
\item{\code{style$title}}{style for the caption (default: \code{"background-color:#999999;vertical-align:top;text-align:left;font-weight:bold;"})}
\item{\code{style$row}}{style for the row names (default: \code{"background-color:#999999;vertical-align:top;text-align:left;font-weight:bold;"})}
\item{\code{style$col}}{style for the col names (default: \code{"background-color:#999999;vertical-align:top;text-align:right;font-weight:bold;"})}
\item{\code{style$cell}}{style for the col names (default: \code{c("background-color:#CCCCCC; vertical-align:top; text-align:right;", "background-color:#FFFFFF; vertical-align:top; text-align:right;")})}
\item{\code{style$logical}}{style for a logical matrix entry (default: \code{c("background-color:#CCCCCC; vertical-align:top; text-align:right;", "background-color:#FFFFFF; vertical-align:top; text-align:right;")})}
\item{\code{style$numeric}}{style for a numeric matrix entry (default: \code{c("background-color:#CCCCCC; vertical-align:top; text-align:right;", "background-color:#FFFFFF; vertical-align:top; text-align:right;")})}
\item{\code{style$char}}{style for a character matrix entry (default: \code{c("background-color:#CCCCCC; vertical-align:top; text-align:right;", "background-color:#FFFFFF; vertical-align:top; text-align:left;"})}
\item{\code{format$title}}{\code{fmt} parameter to format the title via \link[base:sprintf]{base::sprintf} (default: \code{"\%s"})}
\item{\code{format$row}}{\code{fmt} parameter to format the row names via \link[base:sprintf]{base::sprintf} (default: \code{"\%s"})}
\item{\code{format$col}}{\code{fmt} parameter to format the col names via \link[base:sprintf]{base::sprintf} (default: \code{"\%s"})}
\item{\code{format$cell}}{\code{fmt} parameter to format a matrix entry via \link[base:sprintf]{base::sprintf}}
\item{\code{format$logical}}{\code{fmt} parameter to format a logical matrix entry via \link[base:sprintf]{base::sprintf} (default: \code{"\%d"})}
\item{\code{format$numeric}}{\code{fmt} parameter to format a numeric matrix entry via \link[base:sprintf]{base::sprintf} (default: \code{"\%f"})}
\item{\code{format$char}}{\code{fmt} parameter to format a character matrix entry via \link[base:sprintf]{base::sprintf} (default: \code{"\%s"})}
}
}
\examples{
x <- matrix(runif(12), ncol=3)
html_matrix(x)
}
