% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypothesis_latex.R
\name{hypothesis_latex}
\alias{hypothesis_latex}
\title{hypothesis_latex}
\usage{
hypothesis_latex(
  left,
  alternative = c("two.sided", "less", "greater"),
  null = NULL,
  right = paste0(left, "_0")
)
}
\arguments{
\item{left}{character: symbol, for example \code{"\\mu"} or \code{"\\pi"}}

\item{alternative}{character: alternative hypotheses}

\item{null}{character: null hypothese (default: \code{NULL})}

\item{right}{character: symbol (default: \code{paste0(left, "_0")})}
}
\value{
data frame with hypothesis pairs
}
\description{
Creates a data frame with two columns \code{H0} and \code{H1} containing latex representations of the null and the alternative 
hypothesis of a test. If \code{null} is not given then it determined from \code{alternative}. The terms 
\code{"two.sided"},  \code{"greater"}, and  \code{"less"} for \code{alternative} and \code{null} are interpreted 
as in \code{t.test}.
}
\details{
The terms \code{"lt"},  \code{"le"}, \code{"eq"},  \code{"ne"}, \code{"ge"}, and \code{"gt"} are interpreted as is. 
Note that abbreviations, for example \code{alternative="t"} for a two sided test, can not be used.
}
\examples{
hypothesis_latex("\\\\mu")
hypothesis_latex("\\\\pi")
hypothesis_latex("\\\\mu", alternative="two.sided")
hypothesis_latex("\\\\mu", alternative="two.sided", null="lt")
hypothesis_latex("\\\\mu", alternative="ne", null="eq")
}
