% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodle_m2s.R
\name{moodle_m2s}
\alias{moodle_m2s}
\title{moodle_m2s}
\usage{
moodle_m2s(file, newfile = file, quiet = TRUE)
}
\arguments{
\item{file}{character: Moodle XML file with exercises to read from}

\item{newfile}{character:  Moodle XML file to write to (default: \code{file})}

\item{quiet}{logical: generate output during reading (default: \code{FALSE})}
}
\value{
invisibly the file name written to
}
\description{
The \code{rexams} package does not support multiple choice questions with several correct answers but only one allowed answers.
The function reads in the Moodle XML file and replaces \strong{all} \code{<single>false</single>} by \code{<single>true</single>}. Also,
\strong{all} fractions are changed, anything below 0 is set to zero and anything above 0 is set to 100. At the end the modified
XML code is saved into \code{newfile}.
}
\examples{
x <- 1
}
